+++ crates/matrix-sdk-ui/src/timeline/sliding_sync_ext.rs
@@ -34,7 +34,9 @@ impl SlidingSyncRoomExt for SlidingSyncRoom {
     /// EventTimelineItem.
     #[instrument(skip_all)]
     async fn latest_timeline_item(&self) -> Option<EventTimelineItem> {
-        let latest_event = self.latest_event()?;
+        let latest_event =
+            self.client().get_room(self.room_id()).and_then(|room| room.latest_event())?;
+
         EventTimelineItem::from_latest_event(self.client(), self.room_id(), latest_event).await
     }
 }
@@ -62,7 +64,7 @@ mod tests {
         // Given a room with no latest event
         let room_id = room_id!("!r:x.uk").to_owned();
         let client = logged_in_client(None).await;
-        let room = SlidingSyncRoom::new(client, room_id, v4::SlidingSyncRoom::new(), Vec::new());
+        let room = SlidingSyncRoom::new(client, room_id, None, Vec::new());
 
         // When we ask for the latest event, it is None
         assert!(room.latest_timeline_item().await.is_none());
@@ -78,12 +80,7 @@ mod tests {
         process_event_via_sync_test_helper(room_id, event, &client).await;
 
         // When we ask for the latest event in the room
-        let room = SlidingSyncRoom::new(
-            client.clone(),
-            room_id.to_owned(),
-            v4::SlidingSyncRoom::new(),
-            Vec::new(),
-        );
+        let room = SlidingSyncRoom::new(client.clone(), room_id.to_owned(), None, Vec::new());
         let actual = room.latest_timeline_item().await.unwrap();
 
         // Then it is wrapped as an EventTimelineItem